adjustments to user routes

+++ src/services/user/deleteUser.service.ts
@@ -4,7 +4,8 @@ import { AppError } from "../../errors/AppError";
 
 
 const deleteUserService = async (
-  id: string
+  id: string,
+  userId: string
 ): Promise<void> => {
   const userRepository = AppDataSource.getRepository(User);
 
@@ -12,6 +13,10 @@ const deleteUserService = async (
 
   if(!findUser) {
     throw new AppError(404, "User Not Found");
+  }
+
+  if(userId != findUser.id) {
+    throw new AppError(403, "You cannot delete other users")
   }
 
   await userRepository.delete({id})
